requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-jose
passlib
bcrypt
python-multipart
deta
itsdangerous